344 fixed timer bug
