+++ mob_test.go
@@ -1723,37 +1723,6 @@ func TestHelpRequested(t *testing.T) {
 	equals(t, true, helpRequested([]string{"s", "10", "-h"}))
 }
 
-func TestOpenTimerInBrowserWithTimerRoom(t *testing.T) {
-	mockOpenInBrowser()
-	output, configuration := setup(t)
-	configuration.TimerRoom = "testroom"
-
-	err := openTimerInBrowser(configuration)
-
-	assertOutputNotContains(t, output, "Timer Room is not configured.")
-	assertNoError(t, err)
-}
-
-func TestOpenTimerInBrowserWithoutTimerRoom(t *testing.T) {
-	mockOpenInBrowser()
-	output, configuration := setup(t)
-
-	err := openTimerInBrowser(configuration)
-
-	assertOutputContains(t, output, "Timer Room is not configured.")
-	assertNoError(t, err)
-}
-
-func TestOpenTimerInBrowserError(t *testing.T) {
-	mockOpenInBrowser()
-	_, configuration := setup(t)
-	configuration.TimerUrl = ""
-
-	err := openTimerInBrowser(configuration)
-
-	assertError(t, err, "Timer url is not configured")
-}
-
 func TestMobClean(t *testing.T) {
 	_, configuration := setup(t)
